requirements.txt
flask>=3.0.0
playwright>=1.42.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
gunicorn>=21.2.0
